+++ AntraDalis/src/psp/antra/uzduotis/EmailModel.java
@@ -1,4 +1,4 @@
-package psp.uzduotis.antra;
+package psp.antra.uzduotis;
 
 class EmailModel {
     String localPart;
