+++ AntraDalis/src/psp/antra/uzduotis/PhoneValidationRule.java
@@ -1,4 +1,4 @@
-package psp.uzduotis.antra;
+package psp.antra.uzduotis;
 
 class PhoneValidationRule {
     String countryCode;
