+++ AntraDalis/src/psp/antra/uzduotis/PhoneValidator.java
@@ -1,4 +1,4 @@
-package psp.uzduotis.antra;
+package psp.antra.uzduotis;
 
 public class PhoneValidator {
