+++ AntraDalis/test/psp/antra/uzduotis/PhoneValidatorTest.java
@@ -1,10 +1,9 @@
-package psp.uzduotis.antra;
+package psp.antra.uzduotis;
 
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 
 class PhoneValidatorTest {
 
@@ -27,18 +26,18 @@ class PhoneValidatorTest {
         assertEquals("The phone prefix is correct", phoneValidator.prefixRequirement("LT", "+37061234578"));
     }
 
-//    @Test
-//    void PhoneValidator_OtherCountryPhoneNumberLengthCorrect(){
-//        assertEquals("The phone number length is correct", phoneValidator.prefixRequirement("LT", "+37061234578"));
-//    }
+    @Test
+    void PhoneValidator_OtherCountryPhoneNumberLengthCorrect(){
+        assertEquals("The phone number length is correct", phoneValidator.lengthRequirement("LT", "+37061234578"));
+    }
 
     @Test
     void PhoneValidator_OtherCountryPhoneNumberPrefixWrong(){
         assertEquals("The phone prefix does not match country code", phoneValidator.prefixRequirement("PL", "+37061234578"));
     }
 
-//    @Test
-//    void PhoneValidator_OtherCountryPhoneNumberLengthWrong(){
-//        assertEquals("The phone number is too long", phoneValidator.prefixRequirement("PL", "+370612345787878787"));
-//    }
+    @Test
+    void PhoneValidator_OtherCountryPhoneNumberLengthWrong(){
+        assertEquals("The phone number is too long", phoneValidator.lengthRequirement("PL", "+370612345787878787"));
+    }
 }
